link the js file started building the game
created a variable for the maximum number to be chosen within a range (for exemple, user selects 10, so it randomises numbers from 1-10). created a constant to find the targeted number, setting the minimum, adjusting with a +1 so it counts starting from index 1 and, of course, using the random method to generate a random number.