the actual game
created a nem variable, parsing the prompt so it isn't incompatibility with the prompt string and the numeric variable. created a while loop so next prompts are on a loop in case a) guess is greater than the target number, or else the guess is too low.  added a console.log just to properly confirm when then guesses hit the target number.