console.log message
updated the console.log message to actually display the number of attempts. to do that, created a new variable with the name attempts, set it to start at 1, then added it to the while loop so every time the loop begins, one new attempt is added. wrote the message in a string template literal to make it easier to add in the attempts variable.