Update app.js
moved the parse from both if and if else, so there's a work around to quit the game. if parseInt were kept where they were, they would ignore the letter values and keep a while loop until the target number was correctly guesses. the loop itself was parsed, so when the variable guess gets the input value of 'quit', it stops running the loop code with the break keyword.